Phone Input Bug Fix ( 12120 )
`PhoneInput` component will mishandle certain U.S. phone numbers with no leading +1, specifically, any phone numbers where the area code has a 1 as its middle digit (e.g. 714, 213, etc.).  The root cause seems to be in the input masking (i.e. the mask regular expression is looking for a second digit of 1, and lo and behold, find it, and proceeds to "do the wrong thing" with it).  This fix will prepend leading + and country dial code to all inputs that don't have it, resolving the issue.  It _seems_ safe ... i.e. hopefully it is safe to assume that the input will always either start with + and dial code, or else it will exclude them (and just contain the "national" portion of the phone number).