fix: _really_ fix failing unit test (I mean it this time ;-)
Sorry about that.  For some reason, I cannot run the React CM UI unit tests locally until #143 is merged :-(  I was able to create a temp local branch based off of the source branch for that PR as well as this feature branch to verify this version does indeed pass.