fix: `data-testid` issue for `<DataGroups>`
This fixes the following warning (and makes the `data-testid` actually useful!)
```
react_devtools_backend.js:4061 Warning: React does not recognize the `dataTestId` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `datatestid` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
    in div (created by DataGroups)
    in DataGroups (created by PageDataGroups)
    in PageDataGroups (created by [Some Component Using Page Data Groups])
    ...
```
Also adds a unit test fixture for `<DataGroups>` and fixes a typo I noticed in `<DemoPage>` in the docs.