fix: bug fix in `<ActionBarSearch>`
This is a minor regression with the newer functionality in `<ActionBarSearch>` where a `<Select>` con be combined with the search input, in order to help give context to a search term (to aid certain back-end calls that require this to correctly process searches where we're not just using a fell-text search powered by ElasticSearch), introduced in #452.  Some validation logic resulted in a console warning about not using this new feature correctly, even if _consumer was not using it at all!_  This change corrects that issue.